requirements.txt
numpy==1.22.0
Pillow==9.2.0
PyMuPDF==1.20.1
pyqtgraph==0.12.4
PyQt5==5.15.7
setuptools==56.0.0
trimesh==3.12.9
PyOpenGL==3.1.6
